fix disable interrupt

+++ swapper/sleep.cpp
@@ -15,7 +15,7 @@ std::shared_ptr<spdlog::logger> sleepConsole = spdlog::stdout_color_mt("sleep");
 void checkSleep(SwapperManager *manager) {
     sleepLock.lock();
     
-    unsigned long ns = timeSpec.tv_sec * 1000000000ULL + timeSpec.tv_nsec;
+    unsigned long ns = (uint64_t)timeSpec.tv_sec * 1000000000ULL + (uint64_t)timeSpec.tv_nsec;
     clock_gettime(CLOCK_MONOTONIC, &timeSpec);
     auto it = sleepMap.begin();
     while (it != sleepMap.end() && it->first < ns) {
@@ -28,9 +28,11 @@ void checkSleep(SwapperManager *manager) {
 }
 
 void sleepRequestHandler(SleepRequest *req) {
+    //int counter = 0xfffff;
+    //while (counter --) {}
     sleepLock.lock();
     unsigned long target = req->ns + 
-        timeSpec.tv_sec * 1000000000ULL + timeSpec.tv_nsec;
+        (uint64_t)timeSpec.tv_sec * 1000000000ULL + (uint64_t)timeSpec.tv_nsec;
     //sleepConsole->info("sleeping for {}", req->ns);
     sleepMap.insert(std::make_pair(target, req));
